Rename cli_common.py to cli_options.py and move run/validate logic out
cli_common was misleading: sigridci_feedback.py (on-premise) never used
it, so it wasn't actually common to all three entry points. Split it
into addSigridConnectionArguments (the options genuinely shared by all
three) and addPublishArguments (publish/CI-specific), and moved
runSigridCi/validateOptions into sigridci_runner.py, renaming
runSigridCi to runAnalysis, since they run the analysis rather than
parse CLI options.
